backend/requirements.txt
python-can>=4.0
websockets>=12.0
zlgcan>=0.2.0
